README.md
Movie Time is an OOP2 and DB project 
Our application main concept is publish a cinema application for the community. 
We wanted to design this program to help the cinemas customers gain a better experience in booking a ticket and browsing for movies information and cancel a ticket if they can’t make it .
Also the customers can choose which seat they want to choose so they can have the best view. Lastly it’s helpful for the cinema mangers to have an overall presentation of the customers.

programming languages using OOP2 -Java and DataBase